refactor(operator): isolate legacy DGDR read compatibility
